BUGFIX ::: Fixing sonarqube vulnerabilities
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@44556 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 4046f8f39531dbece6a89674a22720690705c0ce

+++ src/prerna/auth/User.java
@@ -722,7 +722,7 @@ public class User implements Serializable {
 							jepThread = PyUtils.getInstance().getJep();
 							this.pyt = new PyTranslator();
 							this.pyt.setPy(jepThread);
-							int logSleeper = 1;
+							long logSleeper = 1;
 							while(!jepThread.isReady())
 							{
 								try
